==> notebooks/02_data_imputation_model_build.ipynb <==
# Used to examine and develop methods for feature imputation

import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LinearRegression
from sklearn.metrics import accuracy_score
from sklearn.impute import SimpleImputer
from sklearn.metrics import mean_squared_error, r2_score
from math import sqrt
from random import randint
import pickle

data_path = '/home/jovyan/work/data/train.csv'
df = pd.read_csv(data_path)

df.head()

df.info()

df.isnull().any()

### First Build columns to impute the Age

gender = {
    'male': 1,
    'female':0
}

embark = {
    'S': 1,
    'C':2,
    'Q':3
}

df['Sex'] = df['Sex'].apply(lambda x: gender.get(x))
df['Embarked'] = df['Embarked'].apply(lambda x: embark.get(x))

df.head()

<strong> As a baseline for performance for assessing the age imputation methods I am going to impute a random value between the min and max. </strong>

def rand_age(cell_value, age_min, age_max):
    if np.isnan(cell_value) == True:
        return randint(age_min, age_max)
    else:
        return cell_value

age_min = round(df['Age'].min())
age_max = df['Age'].max()
df['imp_age_rand'] = df['Age'].apply(rand_age, args=(age_min, age_max))

<strong> I am going to build two new columns to impute the median and mean age. </strong>

df['imp_age_mean'] = df['Age']
df['imp_age_median'] = df['Age']
df['imp_age_median'].fillna((df['imp_age_mean'].median()), inplace=True)
df['imp_age_mean'].fillna((df['imp_age_mean'].mean()), inplace=True)

##  Train model to impute age both Deterministic and Stochastic linear models

df2 = df.copy()
drop_list = ['PassengerId', 'Name', 'Ticket', 'Cabin', 'imp_age_rand', 'imp_age_mean', 'imp_age_median']
df2 = df2.drop(drop_list, 1)
df2.dropna(inplace=True)
df2.head()

df2.isnull().any()

#Create the test and train data sets
x = df2.drop(['Age', 'Survived'], axis=1)
y = df2['Age']
x_train, x_test, y_train, y_test = train_test_split(x, y, test_size=.25, random_state=42)

<strong> Build the Deterministic Linear Model First </strong>

lin_model = LinearRegression()
lin_model.fit(x_train, y_train)

y_pred = lin_model.predict(x_test)
# The coefficients
print('Coefficients: \n', lin_model.coef_)
# The mean squared error
print('Mean squared error: %.2f'
      % mean_squared_error(y_test, y_pred))
# The root mean squared error
rms = sqrt(mean_squared_error(y_test, y_pred))
print(f'The Root Mean squared error: {rms}')
# The coefficient of determination: 1 is perfect prediction
print('Coefficient of determination: %.2f'
      % r2_score(y_test, y_pred))

#Save the Model
filename = '/home/jovyan/work/src/app/assets/models/data_preprocessing_models/age_lin_model_11_30_2020.sav'
pickle.dump(lin_model, open(filename, 'wb'))

<strong> Now add a new column for the Deterministic Linear Regression </strong>

df2.head()

def lin_model_age(df):
    if np.isnan(df['Age']) == True:
        xnew = [[df['Pclass'], df['Sex'], df['SibSp'],df['Parch'], df['Fare'], df['Embarked']]]
        pred_age = lin_model.predict(xnew)[0]
        return pred_age
    else:
        return df['Age']

#Note df2 was used to train the model, we are now applying this model to df1
df['imp_age_det'] = df.apply(lin_model_age, axis=1)

<strong> Now to add some randomness to the linear age regression we will use a stochastic method. </strong>

def sto_lin_model_age(df, std_error):
    if np.isnan(df['Age']) == True:
        xnew = [[df['Pclass'], df['Sex'], df['SibSp'],df['Parch'], df['Fare'], df['Embarked']]]
        pred_age = lin_model.predict(xnew)[0]
        sto_age = np.random.normal(loc=pred_age, scale=std_error)
        return(sto_age)
    
    else:
        return df['Age']

std_error = df['Age'].sem(axis=0, )
std_error = [getattr(std_error, "tolist", lambda: std_error)()]
df['imp_age_det'] = df.apply(sto_lin_model_age, args= (std_error), axis=1)

df.head(10)

df.isnull().any()

### With the imputation models complete and saved these will be used in the data_preprocessing_final notebook to train the primary prediction models
