==> notebooks/05_kaggle_validation.ipynb <==
# The notebook is to run the Kaggle Validation dataset through the prediction API.  This is used to test both the single and multiple call endpoints.

import pandas as pd
import requests
import datetime as dt
import json

#Load in Kaggle validation data
data_path = '/home/jovyan/work/data/test.csv'
kaggle_df = pd.read_csv(data_path, index_col=0)

## First test the single call by using an apply

turl = 'http://192.168.86.44:8002/singletitanicpred'

def single_test(df, turl):
    my_values = {"pclass": df['Pclass'],"sex": df['Sex'],"sibsp": df['SibSp'],"parch": df['Parch'],"fare": df['Fare'],"embarked": df['Embarked'],"age": df['Age'],"passid": df['PassengerId']}
    pred_responce = requests.post(turl, json = my_values)
    pred_responce_json = pred_responce.json()
    return(pred_responce_json['prod_prediciton_binary'])

# Currently FastAPI does not handel NaN values.  While there is server side validation, if I pass a NaN FastAPI will not accept it.  As a result I will convert NaN values before I send.
kaggle_df['Fare'] = kaggle_df['Fare'].fillna(-1)
kaggle_df['Embarked'] = kaggle_df['Embarked'].fillna('NA')
kaggle_df['Age'] = kaggle_df['Age'].fillna('-1')
kaggle_df.reset_index(level=0, inplace=True)

tic = dt.datetime.now()                   
kaggle_df['Prediction'] = kaggle_df.apply(single_test, args=(turl,), axis=1)
toc = dt.datetime.now()
print(f"Total Time took {toc - tic}")

kaggle_df.head(10)

### Total Time was 4 seconds to call each row.  This slow responce is due to the numerous API calls.  It would be faster to write one API call and send all the data at once

#Copy DataFrame and pre
tic = dt.datetime.now() 
tempcopy = kaggle_df.copy()
#Prepare data for sending to URL
tempcopy.reset_index(level=0, inplace=True)
apidf = tempcopy[['Pclass', 'Sex', 'SibSp', 'Parch', 'Fare', 'Embarked', 'Age', 'PassengerId']].copy()
apidf.rename({"Pclass": "pclass", "Sex": "sex", 'SibSp': 'sibsp', 'Parch':'parch', 'Fare': 'fare', 'Embarked':'embarked', 'Age':'age', 'PassengerId': 'passid'}, axis=1, inplace=True)
#Currently FastAPI does not handel NaN values.  While there is server side validation, if I pass a NaN FastAPI will not accept it.  As a result I will convert NaN values before I send.
apidf['fare'] = apidf['fare'].fillna(-1)
apidf['embarked'] = apidf['embarked'].fillna('NA')
apidf['age'] = apidf['age'].fillna('-1')
apidf_json = apidf.to_json(orient = 'records')
# Send data to URL
xurl = 'http://192.168.86.44:8002/multititanicpred' # API URL
xurl_responce = requests.post(xurl, data = apidf_json)
results = xurl_responce.json()
results = results['results']
#Turn Results into dataframe
responce_df = pd.DataFrame(results)
responce_df.rename({'passid': 'PassengerId'}, axis=1, inplace=True)
responce_df.set_index("PassengerId", inplace=True)
#Merge with orginal results
newdf = kaggle_df.join(responce_df)
toc = dt.datetime.now()
print(f"Total Time took {toc - tic}")
newdf.head()

### By passing a whole list instead of one at a time the time to results took less than 1/2 second.
